refactoring in response to PR feedback
